Make makefile aware of OS & ARCH
Now the output files will be in a folder called win, linux, or mac based on the OS environment. Also, the architecture (e.g. x86_64) will be appended to that folder name as well.

If the OS environment is not supported (e.g. Cygwin, Solaris, etc...) the makefile will error.

Also, -DLINUX, -DMACOS, and -DWIN32 are explicitly added based on the OS detected, and on Windows we statically link dlls so that the user can run the .exe without mingw.